Convert options menus to disable links rather than hide them

+++ scripts/ui/mainPageMenuHandler.js
@@ -3,13 +3,25 @@ define(["textResources", "_OptionMenu", "jquery"], function (resources, _OptionM
     var methods = {};
 		
 	/*
-	 * @manager:          The pedalBoardManager.js object to manage pedal boards with
-	 * @logger:           The change logger instance used with the manager
-	 * @undoer:           An undoHandler instance
-	 * @openTutorial:     Function that should open the tutorial popup (or close it if it is open)
-	 * @openHistoryPopup: Function that should open the historyPopup (or close it if it is open)
+	 * Handles click events to the main page menu icon
+	 * 
+	 * PARAMS:
+	 *   @manager:          (PedalBoardManager) The pedalBoardManager instance to manage pedal boards with
+	 *   @logger:           (ChangeLogger)      The change logger instance used with the manager
+	 *   @undoer:           (UndoHandler)       An undoHandler instance
+	 *   @openTutorial:     (function)          Calling this should open the tutorial popup (or close it if it is open)
+	 *   @openHistoryPopup: (function)          Calling this should open the history popup (or close it if it is open)
+	 *
+	 * @returns: The loaded menu
 	 */
 	methods.handle = function(manager, logger, undoer, openTutorial, openHistoryPopup) {
+		
+		/* Gather data up front */
+		var isAnyBoards = manager.Any();
+		var canUndo = undoer.canUndo();
+		var canRedo = undoer.canRedo();
+		
+		/* Start creating the button elements */
 		var addBoardButton = $("<div>")
 			.text(resources.addPedalBoardButtonText)
 			.click(function () {
@@ -41,10 +53,15 @@ define(["textResources", "_OptionMenu", "jquery"], function (resources, _OptionM
 			.append(addBoardButton);
 				
 		var deleteAllBoards = $("<div>")
+			.toggleClass(_OptionMenu.disabledClass, !isAnyBoards)
 			.text(resources.clearAllBoards)
 			.click(function () {
-				if (confirm(resources.clearAllBoardsConfirm))
-					manager.DeleteAll();
+				if (isAnyBoards) {
+					if (confirm(resources.clearAllBoardsConfirm))
+						manager.DeleteAll();
+				} else {
+					_OptionMenu.message(resources.thereAreNoBoardsToDelete);
+				}
 			});
 
 		var deleteAllSection = $("<div>", { "class": "section" })
@@ -54,35 +71,37 @@ define(["textResources", "_OptionMenu", "jquery"], function (resources, _OptionM
 			.text(resources.historyPopupTitle)
 			.click(openHistoryPopup);
 		
-		var historySection = $("<div>", { "class": "section" })
-			.append(historyButton);
-		
 		var undoButton = $("<div>")
+			.toggleClass(_OptionMenu.disabledClass, !canUndo)
 			.text(resources.undoLastChange)
-			.click(undoer.undo);
+			.click(function () {
+				if (canUndo)
+					undoer.undo();
+				else
+					_OptionMenu.message(resources.noChangesToUndo);
+			});
 			
 		var redoButton = $("<div>")
+			.toggleClass(_OptionMenu.disabledClass, !canRedo)
 			.text(resources.redoLastUndo)
-			.click(undoer.redo);
-		
-		var useUndo = undoer.canUndo();
-		var useRedo = undoer.canRedo();
-		if (useRedo)
-			historySection.prepend(redoButton);
-		if (useUndo)
-			historySection.prepend(undoButton);
+			.click(function () {
+				if (canRedo)
+					undoer.redo();
+				else
+					_OptionMenu.message(resources.noUndoneChangesToRedo);
+			});
+					
+		var historySection = $("<div>", { "class": "section" })
+			.append(undoButton)
+			.append(redoButton)
+			.append(historyButton);
 
-		var menuOptions = (useUndo || useRedo)
-			? addBoardSection.add(historySection)
-			: addBoardButton.add(historyButton);
+		var menuOptions = addBoardSection
+			.add(historySection)
+			.add(deleteAllSection);
 		
-		if (manager.Any())
-			menuOptions = menuOptions.add((useUndo || useRedo)
-				? deleteAllSection
-				: deleteAllBoards);
-		
-
-		_OptionMenu.create(menuOptions).addClass("main-page-menu fixed");
+		return _OptionMenu.create(menuOptions)
+			.addClass("main-page-menu fixed");
 	};
 		
     return methods;
